feat(ast): attribute inside a prolog are now of XMLPrologAttribute type
The parent property of an `XMLAttribute` cannot be defined as type "XMLElement".
That is because a prolog may also contain attributes.

To enable consistent handling of the vastly more common use case of
XMLElements being the parents of XMLAttributes, Prolog attributes are now
marked with a special type "XMLPrologAttribute" with a separate interface
and a separate visitor method.

BREAKING CHANGE: Vistors that need to handle attributes inside prolog must now implement the `visitPrologAttribute` method.